Set up confirm-email endpoint in AuthController
The confirm endpoint accepts userId and emailToken as parameters and uses Identity's ConfirmEmailAsync function to confirm the email. The function checks for userId and token being null, whether getting user can be checked and whether email is already confirmed. Sends new EmailConfirmationDto as a response, which includes a Message to send to the frontend and the bool EmailConfirmed.