Set up functionality resend confirmation email in Protected Route
The function handleClick now tries to resend the confirmation email, and if it succeeds, removes the banner with the button and instead only shows the banner for 'Please confirm email'. In order to do this successfully, setUser is exposed from the AuthContext and the response from the function postApiAuthResendConfirmationEmail() populates user with its response.

Since it is a transient value, it only needs to live in the AuthResponseDto on the frontend and not in the User model.