Introduces env.IsDevelopment() to ResendEmailSender
This is done in order to send all confirmation emails to Resends 'delivered@resend.dev' during development, as it is very picky about which emails get to be used. This way we can set which email we like for the user, but the email is not sent to that email. We can still see the send and click the confirmation link in the Resend Emails UI.