Sets up hasUnSavedChanges state
This tracks whether the user has unsaved changes. This will be used in order to show a modal when the user tries to navigate away.

This also sets an setsCustomDesignInternal which we then wrap with our own setCustomDesign wrapper, which sets the hasUnSavedChanges to true when setCustomDesignInternal is run. There is also a check to make sure the flag is not set when the design is loaded from the db.