Sets up conditional rendering in StructuralForm
Introduces a state called showFormElement which has the union type: <"measurements | "wall-color" | "flooring" | null>.

When buttons are clicked for the relevant element of the form, this is either shown or hidden depending on the value of the state.

Changed <form> to <div> to prevent having to deal with defaul behavior from form. Could be changed with minor changes to code, but then we have to set up event.preventDefault() in the onClick functions.